Refactor GraphicsEngine update into two steps: update and paint
We need to separate out the process of updating the graphics to allow things like AnimatedSprites or anything similar to update their internal states based on the time since they were last called. This means that some classes, like TShape and its children, will do nothing when update is called on them.