Implement a WIP TLabel object, requires manually calculating coords
We need to be able to query the current Graphics context for the FontRenderContext in order to be able to calculate the properties of the label (width, height, ascent, baseline, etc) properly. Until we implement this, we have to just eyeball where to place text and hardcode it - we can't ask the label for its width or height in order to exactly place it.